pyproject.toml
[project]
name = "default"
version = "0.1.0"
description = "This repository contains configurations to set up a Python development environment using VSCode's Dev Container feature. The environment includes uv and Ruff."

requires-python = ">=3.10"
readme = "README.md"
license = { file = "LICENSE" }

dependencies = [
    "anyio>=4.10.0",
    "eth-account>=0.13.7",
    "golem-base-sdk==0.1.0",
    "pydantic[email]>=2.10.6,<3.0.0",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "pytz>=2024.2,<2025.0",
]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.uv]
dev-dependencies = [
    "default",
    "pyright>=1.1.389,<2.0.0",
    "pytest>=8.3.3,<9.0.0",
    "ruff>=0.8.0,<0.9.0",
]

[tool.uv.sources]
default = { workspace = true }

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",
    "COM819",
    "D100",
    "D203",
    "D213",
    "D300",
    "E111",
    "E114",
    "E117",
    "ISC001",
    "ISC002",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "T201",
    "W191",
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",
    "ARG",
    "FBT",
    "PLR2004",
    "S311",
]

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "All"
venvPath = "."
venv = ".venv"
include = ["src"]
exclude = [
    "**/__pycache__",
    ".pytest_cache",
    ".ruff_cache",
    ".venv"
]
